\name{CreateBinary}
\alias{CreateBinary}
\encoding{UTF-8}
\title{Create binary file}
\description{
  Matrix decomposition.
}
\usage{
CreateBinary(dev,fname)
}
\arguments{
  \item{dev}{numeric; indicates de dev ID}
  \item{fname}{string; the name of the function}
}

\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}



\keyword{STBEU}
