\name{print.STBEUFit}
\alias{print.STBEUFit}
\encoding{UTF-8}
\title{Matrix decomposition}
\description{
  Matrix decomposition.
}
\usage{
print.STBEUFit(x,names,GPU,varest, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{A}
  \item{names}{B}
  \item{GPU}{C}
  \item{varest}{D}
  \item{digits}{D}
  \item{...}{G}
}

\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}



\keyword{STBEU}
