\name{setting_param}
\alias{setting_param}
\encoding{UTF-8}
\title{Checking Correlation Model}
\description{
  The procedure controls if the correlation model
  inserted is correct.
}
\usage{
setting_param(cc,theta,fix)
}
\arguments{
  \item{fix}{Numeric; a vector which has fixed parameters info.}
  \item{theta}{Numeric; a vector which has to be estimated parameters info.}
  \item{cc}{Numeric; it indicates the correlation model}
}

\value{Return the \code{theta} parameters with correct names.} 




\author{Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},\url{https://sites.google.com/a/uv.cl/moreno-bevilacqua/home},
Víctor Morales Oñate, \email{victor.morales@uv.cl}, \url{https://sites.google.com/site/moralesonatevictor/}
}

\keyword{Composite}
